Add dynamic show title
