+++ src/pages/Show/Show.jsx
@@ -1,9 +1,18 @@
 import React from "react";
 import { useParams } from "react-router-dom";
 
-const Show = () => {
+const Show = ({ showsData }) => {
   const { id } = useParams();
-  return <div>Show {id}</div>;
+
+  if (!showsData[id]) {
+    return <div>Loading...</div>;
+  }
+
+  return (
+    <div>
+      <h1>{showsData[id].show.name}</h1>
+    </div>
+  );
 };
 
 export default Show;
